chore(sweep): move javm-exec tests out of src/ (10 of 11 modules)
Continuing the sweep. javm-exec is large (11 inline `mod tests`
blocks); 10 only touch the crate's public API and move cleanly to
`javm-exec/tests/`:

- args, decode, ecall, gas, gas_sim, instruction, interp, mem,
  program, regs

`gas_cost` stays inline — its `mod proptests` exercises private
types (`RegSet`, `ExecUnits`, `reg_bit`). That's exactly the case
the inline form is reserved for.

Imports rewritten from `use super::*` / `use crate::foo::Bar` to the
crate-name form (`use javm_exec::*` and friends). Proptest deps
stay; the integration tests pick them up via `dev-dependencies`.

Verified: `cargo test -p javm-exec` green — 13 inline (gas_cost
proptests) + 88 across the new integration test files = 101 tests.
